+++ R/convert.R
@@ -1,77 +1,3 @@
-#' Convert register SAS file(s) and save to Parquet format
-#'
-#' @description
-#' This function reads one or more SAS files for a given register, and saves the
-#' data in Parquet format. It expects the input SAS files to come from the same
-#' register, e.g., different years of the same register. The function checks
-#' that all files belong to the same register by comparing the alphabetic
-#' characters in the file name(s).
-#'
-#' The function looks for a year (1900-2099) in the file
-#' names in `path` to use the year as partition, see `vignette("design")`
-#' for more information about the partitioning.
-#'
-#' If a year is found, the data is saved as a partition by year in the output
-#' directory, e.g., `output_dir/register_name/year=2020/part-ad5b.parquet`
-#' (the ending being a UUID). If no year is found in the file name, the data
-#' is saved in a
-#' `year=__HIVE_DEFAULT_PARTITION__` partition, which is the standard Hive
-#' convention for missing partition values.
-#'
-#' Two columns are added to the output: `source_file` (the original SAS file
-#' path) and `year` (extracted from the file name, used as partition key).
-#'
-#' To be able to handle larger-than-memory SAS files, this function uses
-#' `convert_file()` internally and only converts one file at a time in chunks.
-#' As a result, identical rows are not deduplicated.
-#'
-#' @param path Paths to SAS files for one register. See [list_sas_files()].
-#' @param output_dir Directory to save the Parquet output to. Must not include
-#'  the register name as this will be extracted from `path` to create the
-#'  register folder.
-#' @param chunk_size Number of rows to read and convert at a time.
-#'
-#' @returns `output_dir`, invisibly.
-#'
-#' @export
-#' @examples
-#' sas_file_directory <- fs::path_package("fastreg", "extdata")
-#' convert_register(
-#'   path = list_sas_files(sas_file_directory),
-#'   output_dir = fs::path_temp("path/to/output/register/")
-#' )
-convert_register <- function(
-  path,
-  output_dir,
-  chunk_size = 10000000L
-) {
-  # Check that register dir is empty (if exists) to avoid duplicating data
-  # since parts are named with UUIDs.
-  # Get register name checks that only one register is in `path`.
-  register_dir <- fs::path(output_dir, get_register_name(path))
-  if (fs::dir_exists(register_dir) && length(fs::dir_ls(register_dir)) > 0) {
-    cli::cli_abort(c(
-      "Output directory is not empty: {.path {register_dir}}",
-      "i" = "Delete the directory manually before re-running."
-    ))
-  }
-
-  # Convert files.
-  purrr::walk(
-    path,
-    \(p) convert_file(p, output_dir, chunk_size)
-  )
-
-  # Success message.
-  cli::cli_alert_success("Successfully converted {length(path)} file{?s}.")
-  cli::cli_bullets(c(
-    "*" = "Input: {.val {fs::path_file(path)}}",
-    "*" = "Output: Register files in {.path {fs::path(output_dir, get_register_name(path))}}"
-  ))
-
-  invisible(output_dir)
-}
-
 #' Convert a single register SAS file to Parquet
 #'
 #' To be able to handle larger-than-memory files, the SAS file is converted in
@@ -80,18 +6,21 @@ convert_register <- function(
 #' exists in the directory, since files are saved with UUIDs in their names.
 #'
 #' @param path Path to a single SAS file.
-#' @inheritParams convert_register
+#' @param output_dir Directory to save the Parquet output to. Must not include
+#'  the register name as this will be extracted from `path` to create the
+#'  register folder.
+#' @param chunk_size Number of rows to read and convert at a time.
 #'
-#' @returns `output_dir`, invisibly.
+#' @returns A tibble with a conversion log about each written chunk.
 #'
 #' @export
 #' @examples
 #' sas_file <- fs::path_package("fastreg", "extdata", "test.sas7bdat")
-#' convert_file(
+#' convert(
 #'   path = sas_file,
 #'   output_dir = fs::path_temp("path/to/output/file")
 #' )
-convert_file <- function(
+convert <- function(
   path,
   output_dir,
   chunk_size = 10000000L
@@ -105,7 +34,8 @@ convert_file <- function(
   # Prepare variables used in repeat below.
   partition_path <- create_partition_path(path, output_dir)
   part <- create_part_uuid()
-  skip <- 0L
+  skip <- 0 # Not integer (0L) since they overflow at ~2.1B rows, doubles don't.
+  conversion_log <- tibble::tribble(~register_name, ~input_path, ~output_path, ~row_count, ~schema)
 
   # Read first chunk to establish schema.
   chunk <- read_sas_chunk(path, skip, chunk_size)
@@ -117,14 +47,28 @@ convert_file <- function(
       break
     }
 
+    file_path <- fs::path(
+      partition_path,
+      glue::glue("part-{part}.parquet")
+    )
     chunk |>
       arrow::as_arrow_table(schema = schema) |>
-      arrow::write_parquet(
-        sink = fs::path(
-          partition_path,
-          glue::glue("part-{part}.parquet")
-        )
+      arrow::write_parquet(sink = file_path)
+
+    # Add current chunk's info to list.
+    conversion_log <- dplyr::bind_rows(
+      conversion_log,
+      tibble::tibble(
+        register_name = get_register_name(path),
+        input_path = path,
+        output_path = fs::path(file_path),
+        row_count = nrow(chunk),
+        schema = list(tibble::tibble(
+          column_name = colnames(chunk),
+          data_type = purrr::map_chr(chunk, class)
+        ))
       )
+    )
 
     skip <- skip + nrow(chunk)
     part <- create_part_uuid()
@@ -134,13 +78,13 @@ convert_file <- function(
 
   cli::cli_alert_success("Converted {.path {fs::path_file(path)}}")
 
-  invisible(output_dir)
+  conversion_log
 }
 
 #' Read SAS chunk
 #'
 #' @param skip Number of rows to skip.
-#' @inheritParams convert_file
+#' @inheritParams convert
 #'
 #' @returns A tibble.
 #'
@@ -157,7 +101,7 @@ read_sas_chunk <- function(path, skip, chunk_size) {
 #' Gets the year and register name from the file name in `path` and creates
 #' a partition path `{output_dir}/{register_name}/year={year}/`.
 #'
-#' @inheritParams convert_file
+#' @inheritParams convert
 #'
 #' @returns The partition path.
 #'
@@ -167,7 +111,11 @@ create_partition_path <- function(path, output_dir) {
   year <- get_year_from_filename(path)
   # Following the default `null_fallback` in arrow::hive_partition()
   # https://arrow.apache.org/docs/r/reference/hive_partition.html#arg-null-fallback.
-  year_partition <- if (is.na(year)) "__HIVE_DEFAULT_PARTITION__" else year
+  year_partition <- dplyr::if_else(
+    is.na(year),
+    "__HIVE_DEFAULT_PARTITION__",
+    as.character(year)
+  )
   partition_path <- fs::path(
     output_dir,
     get_register_name(path),